+++ nanodlp_format/benches/benchmark.rs
@@ -4,7 +4,7 @@ use nalgebra::Vector2;
 use nanodlp_format::png::{
     PngEncoder, PngInfo,
     deflate::{Adler32, huffman, lz77_compress},
-    insert_runs,
+    intersperse_runs,
 };
 use std::hint::black_box;
 
@@ -13,15 +13,15 @@ const PLATFORM: Vector2<u32> = Vector2::new(11520, 5120);
 fn criterion_benchmark(c: &mut Criterion) {
     let layer = generate_layer();
 
-    c.bench_function("insert_runs", |b| {
+    c.bench_function("Intersperse Runs", |b| {
         b.iter_batched(
             || layer.clone(),
-            |mut runs| black_box(insert_runs(&mut runs, 0, 11520)),
+            |mut runs| black_box(intersperse_runs(&mut runs, 0, 11520)),
             BatchSize::SmallInput,
         )
     });
 
-    c.bench_function("checksum", |b| {
+    c.bench_function("Checksum", |b| {
         b.iter(|| {
             let mut check = Adler32::new();
             layer.iter().for_each(|run| check.update_run(run));
@@ -30,9 +30,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     });
 
     let mut rgb = layer.clone();
-    insert_runs(&mut rgb, 0, 11520);
+    intersperse_runs(&mut rgb, 0, 11520);
 
-    c.bench_function("tokens", |b| {
+    c.bench_function("Tokens", |b| {
         b.iter_batched(
             || rgb.clone(),
             |rgb| black_box(lz77_compress(rgb.into_iter())),
@@ -41,7 +41,7 @@ fn criterion_benchmark(c: &mut Criterion) {
     });
 
     let tokens = lz77_compress(rgb.into_iter());
-    c.bench_function("huffman", |b| {
+    c.bench_function("Huffman", |b| {
         b.iter(|| {
             let mut out = Vec::new();
             let mut bits = BitVec::new(&mut out, 0);
@@ -50,7 +50,7 @@ fn criterion_benchmark(c: &mut Criterion) {
         });
     });
 
-    c.bench_function("layer_encode", |b| {
+    c.bench_function("Layer Encode", |b| {
         b.iter_batched(
             || layer.clone(),
             |rgb| {