+++ Changelog.md
@@ -21,6 +21,7 @@
 - Custom project and mesh (de)serialization with progress monitoring
 - Added task panel to view the progress of running async tasks
 - Run post processing operations asynchronously
+- Added support for the NanoDLP format (.nanodlp)
 
 ## v0.4.0 &mdash; November 25th, 2025